backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
youtube-transcript-api>=1.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
